use size_t instead of unsigned int to avoid MSVC warnings about converting the former to the latter when compiling with /Wp64
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@50242 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/arrstr.h
@@ -404,7 +404,7 @@ public:
     // default copy constructor is ok
 
     // iteration interface
-    unsigned int GetCount() const { return m_size; }
+    size_t GetCount() const { return m_size; }
     bool IsEmpty() const { return GetCount() == 0; }
     const wxString& operator[] (unsigned int i) const
     {
@@ -429,7 +429,7 @@ private:
     };
 
     wxStringContainerType m_type;
-    unsigned int m_size;
+    size_t m_size;
     union
     {
         const wxString *      ptr;
